+++ testing.py
@@ -1,5 +1,3 @@
-
-
 # Print out text string to console and to ConsoleOutput.txt
 def printOut(text):
     print(text)
@@ -35,7 +33,9 @@ fileName = os.path.join(dateDirectory, "ConsoleOutput.txt")
 consoleOutputFile = open(fileName, "w")
 
 # Load the test sites from GitHub
-with urllib.request.urlopen("https://raw.githubusercontent.com/USGS-WiM/StreamStats-Setup/master/batchTester/testSites.geojson") as url:
+testSitesURL = 'https://raw.githubusercontent.com/USGS-WiM/StreamStats-Setup/master/batchTester/testSites.geojson'
+fakeTestSitesURL = 'https://raw.githubusercontent.com/USGS-WiM/StreamStats-Setup/testAM/batchTester/testSites.geojson'
+with urllib.request.urlopen(fakeTestSitesURL) as url:
     sites = json.loads(url.read().decode())['features']
 
 servers = ["test", "prodweba", "prodwebb"]
@@ -57,6 +57,27 @@ for server in servers:
         if not os.path.exists(basinCharacteristicsDirectory):
             os.makedirs(basinCharacteristicsDirectory)
 
+        # Create a file where basin characteristics comparison will be saved
+        fileName = os.path.join(basinCharacteristicsDirectory, "BasinCharacteristicsComparison.csv")
+        basinCharacteristicsComparisonFile = open(fileName, "w", newline='')
+        basinCharacteristicsComparisonFileWriter = csv.writer(basinCharacteristicsComparisonFile)
+        headerRow = ["Region", "SiteID", "WorkspaceID", "Latitude", "Longitude", "Basin Characteristic", "Computed value", "Known value", "Computed Value Equal to Known Value?"]
+        basinCharacteristicsComparisonFileWriter.writerow(headerRow)
+
+        # Create a file where computed basin characteristics that are different from known values will be saved
+        fileName = os.path.join(basinCharacteristicsDirectory, "BasinCharacteristicsDifferences.csv")
+        basinCharacteristicsDifferenceFile = open(fileName, "w", newline='')
+        basinCharacteristicsDifferenceFileWriter = csv.writer(basinCharacteristicsDifferenceFile)
+        headerRow = ["Region", "SiteID", "WorkspaceID", "Latitude", "Longitude", "Basin Characteristic", "Computed value", "Known value"]
+        basinCharacteristicsDifferenceFileWriter.writerow(headerRow)
+
+        # Create a file where computed basin characteristics that were not compared to known values will be saved
+        fileName = os.path.join(basinCharacteristicsDirectory, "BasinCharacteristicsUncompared.csv")
+        basinCharacteristicsUncomparedFile = open(fileName, "w", newline='')
+        basinCharacteristicsUncomparedFileWriter = csv.writer(basinCharacteristicsUncomparedFile)
+        headerRow = ["Region", "SiteID", "WorkspaceID", "Latitude", "Longitude", "Basin Characteristic", "Computed value"]
+        basinCharacteristicsUncomparedFileWriter.writerow(headerRow)
+
         # Create a folder for Flow Statistics results: Output/Testing-YYYY-MM-DD-HH-MM-SS/[server]/FlowStatistics
         flowStatsDirectory = os.path.join(serverFolderDirectory, r'FlowStatistics')
         if not os.path.exists(flowStatsDirectory):
@@ -65,8 +86,6 @@ for server in servers:
     serverStartTime = datetime.now()
 
     printOut("========== PERFORMING TESTS ON " + server.upper() + " SERVER ==========")
-
-
 
     for site in sites:
         
@@ -93,16 +112,15 @@ for server in servers:
                 workspaceID = json.loads(response.content)['workspaceID']
                 coordinates = json.loads(response.content)['featurecollection'][1]['feature']['features'][0]["geometry"]['coordinates']
                 
-                # TODO change this from "test" delineation directory (and similar below)
                 siteDelineationFileName = os.path.join(delineationDirectory, region + ".txt")
                 siteDelineationFile = open(siteDelineationFileName, "w")
                 siteDelineationFile.write(json.dumps(json.loads(response.content)))
+                siteDelineationFile.close()
 
                 basinDelineationEndTime = datetime.now()
                 basinDelineationTimeElapsed = basinDelineationEndTime - basinDelineationStartTime
-                printOut("Completed successfully.") # Print the time it took to the console + elapsed time
-                printOut("Time elapsed: " + str(basinDelineationTimeElapsed)) # Print the time it took to the console + elapsed time
-                # Print the time it took in the output.txt + elapsed time
+                printOut("Completed successfully.") 
+                printOut("Time elapsed: " + str(basinDelineationTimeElapsed)) 
             except Exception as e:
                 printOut("Failed. Retrying...")
                 # printOut(e)
@@ -125,21 +143,20 @@ for server in servers:
                     computeBasinCharacteristicsString = "https://{}.streamstats.usgs.gov/streamstatsservices/parameters.json?rcode={}&workspaceID={}&includeparameters=true".format(server, region, workspaceID)
                     response = requests.get(computeBasinCharacteristicsString)
                     parameters = json.loads(response.content)['parameters']
-                    
-                    # printOut(parameters)
 
+                    # Check to make sure that values were actually returned from the service
+                    if 'value' not in parameters[0]:
+                        raise Exception
+                    
                     siteBasinCharcteristicsFileName = os.path.join(basinCharacteristicsDirectory, region + ".txt")
                     siteBasinCharcteristicsFile = open(siteBasinCharcteristicsFileName, "w")
-                    # printOut(json.loads(response.content))
                     siteBasinCharcteristicsFile.write(json.dumps(json.loads(response.content)))
-                    # printOut(json.dumps(json.loads(response.content)))
+                    siteBasinCharcteristicsFile.close()
 
                     basinCharacteristicsEndTime = datetime.now()
                     basinCharacteristicsTimeElapsed = basinCharacteristicsEndTime - basinCharacteristicsStartTime
-                    # basinCharacteristicsElapsed = basinCharacteristicsTime.strftime("%Y/%m/%d %H:%M:%S")
                     printOut("Completed successfully.")
                     printOut("Time elapsed: " + str(basinCharacteristicsTimeElapsed))
-                    # Print the time it took in the output.txt + elapsed time
 
                     printOut("COMPARING BASIN CHARACTERISTICS TO KNOWN VALUES:")
                     printOut("Running...")
@@ -149,7 +166,6 @@ for server in servers:
                     knownBasinCharacteristics = site['properties']['testData']
                     for knownBasinCharacteristic in knownBasinCharacteristics:
                         knownBasinCharacteristicsDictionary[knownBasinCharacteristic["Label"]] = knownBasinCharacteristic["Value"]
-                    knownBasinCharacteristicsDictionary["L3_PIEDMNT"] = 50
 
                     numberBasinCharacteristicsNotEqualToKnownValues = 0
                     for parameter in parameters:
@@ -159,8 +175,25 @@ for server in servers:
                                 printOut("Known value = " + str(knownBasinCharacteristicsDictionary[parameter['code']]))
                                 printOut("Computed value = " + str(parameter['value']))
                                 numberBasinCharacteristicsNotEqualToKnownValues += 1
-                        except:
-                            printOut("No known value for " + parameter['code'] + ". Computed value cannot be compared.")
+                                dataRow = [region, siteID, workspaceID, ylocation, xlocation, parameter['code'], parameter['value'], knownBasinCharacteristicsDictionary[parameter['code']]]
+                                basinCharacteristicsDifferenceFileWriter.writerow(dataRow)
+                                basinCharacteristicsDifferenceFile.flush()
+                            dataRow = [region, siteID, workspaceID, ylocation, xlocation, parameter['code'], parameter['value'], knownBasinCharacteristicsDictionary[parameter['code']], str(knownBasinCharacteristicsDictionary[parameter['code']] == parameter['value'])]
+                            basinCharacteristicsComparisonFileWriter.writerow(dataRow)
+                            basinCharacteristicsComparisonFile.flush()
+                        except Exception as e:
+                            try:
+                                printOut("No known value for " + parameter['code'] + ". Computed value cannot be compared.")
+                                dataRow = [region, siteID, workspaceID, ylocation, xlocation, parameter['code'], parameter['value']]
+                                basinCharacteristicsUncomparedFileWriter.writerow(dataRow)
+                                basinCharacteristicsUncomparedFile.flush()
+                                dataRow = [region, siteID, workspaceID, ylocation, xlocation, parameter['code'], parameter['value'], "None", "Not applicable"]
+                                basinCharacteristicsComparisonFileWriter.writerow(dataRow)
+                                basinCharacteristicsComparisonFile.flush()
+                            except Exception as e:
+                                print(e)
+                                break
+
                     if numberBasinCharacteristicsNotEqualToKnownValues == 0:
                         printOut("Completed. All computed values were equal to known values.")
                     elif numberBasinCharacteristicsNotEqualToKnownValues == 1:
@@ -180,9 +213,11 @@ for server in servers:
     serverEndTime = datetime.now()
     serverTimeElapsed = serverEndTime - serverStartTime
     printOut(server.upper() + " server finished testing! Elapsed time for this server: " + str(serverTimeElapsed)) 
-    # outputFile.write(server.capitalize() + "server finished testing! Elapsed time for this server: " + str(serverTimeElapsed)) 
 
 overallEndTime = datetime.now()
 overallTimeElapsed = overallEndTime - overallStartTime
 printOut("Testing complete! Overall elapsed time: " + str(overallTimeElapsed))
 consoleOutputFile.close()
+basinCharacteristicsComparisonFile.close()
+basinCharacteristicsDifferenceFile.close()
+basinCharacteristicsUncomparedFile.close()